refactor: consolidate Fleet into Settings Network tab; remove redundant nav items
Move node identity, registry connect, and bookmarked-nodes UI from DashboardPage
into a new NetworkTab in Settings. Remove /fleet and /browse routes and their page
files. Remove Browse Agents and Fleet entries from the slide panel nav, keeping
only Receipts and All Settings.
