+++ apps/papillon/frontend/src/app.rs
@@ -11,9 +11,7 @@ use crate::bridge;
 use crate::components::setup_wizard::SetupWizard;
 use crate::components::topbar::TopBar;
 use crate::pages::activity::ActivityPage;
-use crate::pages::browse::BrowsePage;
 use crate::pages::canvas::CanvasPage;
-use crate::pages::dashboard::DashboardPage;
 // home.rs removed — canvas at "/" IS the home screen
 use crate::pages::receipts::ReceiptsPage;
 use crate::pages::scenario::ScenarioPage;
@@ -372,8 +370,6 @@ pub fn App() -> impl IntoView {
                         <Route path=path!("/activity") view=ActivityPage />
                         <Route path=path!("/receipts") view=ReceiptsPage />
                         <Route path=path!("/settings") view=SettingsPage />
-                        <Route path=path!("/browse") view=BrowsePage />
-                        <Route path=path!("/fleet") view=DashboardPage />
                     </Routes>
                 </main>
                 <footer class="status-bar app-statusbar">
